+++ src/config/prod.js
@@ -3,11 +3,6 @@ module.exports = {
     username: process.env.DB_USER || 'admin',
     password: process.env.DB_PASS || 'password',
     host: 'gw2armory-prod.cekbcmynaoxp.us-east-1.rds.amazonaws.com',
-    logging: false,
-  },
-
-  fetch: {
-    verbose: false,
   },
 
   web: {