+++ src/lib/models/user.js
@@ -18,7 +18,7 @@ module.exports = function (sequelize, DataTypes) {
       type: DataTypes.STRING,
       required: true,
       field: 'email',
-      // TODO: Remove constraint.
+      // REQUIRES MIGRATION (turning unique off)
       // unique: true,
       allowNull: false,
     },
@@ -35,9 +35,12 @@ module.exports = function (sequelize, DataTypes) {
       allowNull: false,
       defaultValue: false,
     },
+
+    // REQUIRES MIGRATION
     stub: {
       type: DataTypes.BOOLEAN,
       defaultValue: false,
+      allowNull: false,
     },
   }, {
     classMethods: {