+++ src/lib/services/tokens.js
@@ -1,6 +1,10 @@
 module.exports = function readTokens (models) {
   return models.Gw2ApiToken
-    .findAll()
+    .findAll({
+      where: {
+        stub: false,
+      },
+    })
     .then((items) => items.map((item) => ({
       token: item.dataValues.token,
       permissions: item.dataValues.permissions,